Re-enable RPC server
