+++ rpc-server/Cargo.toml
@@ -29,12 +29,17 @@ base64 = "0.10"
 beserial = { path = "../beserial", version = "0.1.0" }
 nimiq-consensus = { path = "../consensus", version = "0.1.0" }
 nimiq-blockchain = { path = "../blockchain", version = "0.1.0" }
+nimiq-blockchain-albatross = { path = "../blockchain-albatross", version = "0.1.0" }
+nimiq-blockchain-base = { path = "../blockchain-base", version = "0.1.0" }
 nimiq-mempool = { path = "../mempool", version = "0.1.0" }
 nimiq-network = { path = "../network", version = "0.1.0" }
 nimiq-hash = { path = "../hash", version = "0.1.0" }
 nimiq-network-primitives = { path = "../network-primitives", version = "0.1.0" }
 nimiq-block = { path = "../primitives/block", version = "0.1.0" }
+nimiq-block-albatross = { path = "../primitives/block-albatross", version = "0.1.0" }
+nimiq-block-base = { path = "../primitives/block-base", version = "0.1.0" }
 nimiq-transaction = { path = "../primitives/transaction", version = "0.1.0" }
 nimiq-keys = { path = "../keys", version = "0.1.0" }
 nimiq-block-production = { path = "../block-production", version = "0.1.0" }
 nimiq-utils = { path = "../utils", version = "0.1.0", features = ["merkle", "time"] }
+nimiq-primitives = { path = "../primitives", version = "0.1.0" }
